Outlines for Limbo Area and Entities
Ship of Theseus ahh program
Improvements all over the place! dimensional window movement has been completely rehauled and is no longer an active update system. theres an outline that appears to show you where its targeting. This makes preventing window overlaps way easier!
Also I've decided that the player should be able to see the level between screens. It should be super simple, and the focus should remain on the screens, but the player should be able to see what theyre getting into. This has also been applied to entities - they will need another texture for outside of the screens as well.
I've had to make a few minor adjustments to the addon ive installed, and am very grateful to godot for allowing this godot is so cool