Dimensional Window Customisability
The windows now are resizable and movable and are saved in the main metadata to be loaded at runtime. Also, I've implemented polygon splitting so that if a polygon entirely contains a window it'll split it into two polygons and run the clip on both of them. This of course involved completely redoing the build_windows function and resaving all the levels i've built so far. ship of theseus continues.