Same-Layer Windows, Spike Flag integration
Entities now have flag functions which is a smidge intimidating. clipping is slower now, but its only done once every time a window is dropped, so that should counteract it a lot. windows are loaded into the editor when loading levels. stamina increase for the player. added get_flag function to flagmanager.