Final Commit Probably
What a world. It's 10:22 at night and I have more dopamine than I've had in the past 6 months.
Not including when i figured out how to do the clipping thing thats still an all-time high.
um i tried to put this "secret message" in the lore text box of the game, but that broke it so im just gonna slam it in here.

woaahhhh secret message.... through the mountain.... secret message..... is this a reference?
Nah. Surely Not. I Would Never Reference Anything.
I'd like to think i've made a pretty compelling game here.
I am very, very proud of what I've done with the help of so many.
So yeah. Shoutout to the world and to humanity and to HRT one day and etc etc etc.

So, actual commit info:
- pause menu makes the main gameplay music quieter.
- main gameplay music added (thanksben)
- music volume now properly sets itself in PersistentData
- music cuts when CASSANDRA is deactivated.
- Audio Bus changes in a few scenes.

yeah. this is it. i dont thihnk theres anything else to do. what a wild ride.